tox testing, add QC
