--- mypy.ini
@@ -1,13 +1,0 @@
-[mypy]
-follow_imports = silent
-strict_optional = True
-warn_redundant_casts = True
-disallow_any_generics = True
-check_untyped_defs = True
-no_implicit_reexport = True
-ignore_missing_imports = True
-
-# for strict mypy: (this is the tricky one :-))
-disallow_untyped_defs = True
-
-files = src
