+++ setup.cfg
@@ -24,7 +24,7 @@ classifiers =
 package_dir=
     =src
 packages = find:
-py_modules = ome_autogen
+; py_modules = ome_autogen
 include_package_data = true
 python_requires = >=3.7
 install_requires =
@@ -36,8 +36,46 @@ install_requires =
 * = *.xsd
 
 [options.extras_require]
-autogen = isort<5.0; black
-
+autogen = isort<5.0; black; autoflake
 
 [options.packages.find]
 where=src
+
+[flake8]
+ignore = W503, C901, E501
+max-line-length = 88
+
+[mypy]
+follow_imports = silent
+strict_optional = True
+warn_redundant_casts = True
+disallow_any_generics = True
+check_untyped_defs = True
+no_implicit_reexport = True
+ignore_missing_imports = True
+files = src
+disallow_untyped_defs = True
+
+[tox:tox]
+envlist = py{37,38}
+toxworkdir=/tmp/.tox
+
+[gh-actions]
+python =
+    3.7: py37
+    3.8: py38, mypy, flake8
+
+[testenv]
+deps =
+    pytest
+    pytest-cov
+commands =
+    pytest -v --cov-report=xml --cov=ome_types --basetemp={envtmpdir} {posargs}
+    
+[testenv:mypy]
+deps = mypy
+commands = mypy
+
+[testenv:lint]
+deps = flake8
+commands = flake8
