+++ src/ome_types/__init__.py
@@ -4,28 +4,27 @@ except ImportError:
     __version__ = "unknown"
 
 import os
-from .schema import to_dict, validate
-from typing import Union, Type
 from pathlib import Path
-
-__all__ = ["to_dict", "validate", "from_xml"]
-
+from typing import Union
 
 try:
     from .model import OME
 except ImportError:
-    print("OME dataclasses not found ... running autogeneration")
-    import runpy
+    raise ImportError(
+        "Could not import 'ome_types.model.OME'.\nIf you are in a dev environment, "
+        "you may need to run 'python -m src.ome_autogen'"
+    )
+from .schema import to_dict, validate
 
-    runpy.run_module("ome_autogen", run_name="__main__")
-    from .model import OME
+
+__all__ = ["to_dict", "validate", "from_xml"]
 
 
-def from_xml(xml: Union[Path, str], OME=OME) -> OME:  # type: ignore
+def from_xml(xml: Union[Path, str]) -> OME:  # type: ignore
     xml = os.fspath(xml)
     d = to_dict(xml)
     for key in list(d.keys()):
         if key.startswith(("xml", "xsi")):
             d.pop(key)
 
-    return OME(**d)
+    return OME(**d)  # type: ignore
