--- testing/conftest.py
@@ -1,7 +1,0 @@
-def pytest_addoption(parser):
-    parser.addoption(
-        "--nogen",
-        default=False,
-        action="store_true",
-        help="run tests on pregenerated model",
-    )
